Added Firebase Core and Auth for User login
added the Firebase SDK and install the cocoa pods for Firebase
to load up basic Firebase Core and Firebase Auth to allow user login
Also added a sign up page for employers.

From now on open the AllTheWage.xcworkspace instead of the AllTheWage.xcodeproj
to allow coocapods to work from now on.